move coveralls module to own file and remove indenting for modules as recommended in the julia manual
